refactor(perceiver): standardize in_shape, out_shape
