+++ test/module/perceiver_io/test_postprocessor.py
@@ -4,24 +4,27 @@ import torch
 
 
 @pytest.mark.parametrize('batch', [2])
-@pytest.mark.parametrize('output_shape', [(3, 9)])
-@pytest.mark.parametrize('proj_dim', [10])
-def test_projection_postprocessor(batch, output_shape, proj_dim):
-    out = torch.rand(batch, *output_shape)
-    module = postprocessor.ProjectionPostprocessor(output_shape, proj_dim)
-    post_out = module(out)
-    assert list(post_out.shape) == [batch, output_shape[0], proj_dim]
-    assert not post_out.isnan().any()
-    post_out.mean().backward()
+@pytest.mark.parametrize('in_shape', [(3, 9)])
+@pytest.mark.parametrize('out_dim', [10])
+def test_projection_postprocessor(batch, in_shape, out_dim):
+    x = torch.rand(batch, *in_shape)
+    module = postprocessor.ProjectionPostprocessor(in_shape, out_dim)
+    out = module(x)
+    assert [in_shape[0], out_dim] == module.out_shape
+    assert list(out.shape) == [batch, *module.out_shape]
+    assert not out.isnan().any()
+    out.mean().backward()
 
 
 @pytest.mark.parametrize('batch', [2])
-@pytest.mark.parametrize('output_shape', [(1, 9)])
-@pytest.mark.parametrize('num_classes', [10])
-def test_classification_postprocessor(batch, output_shape, num_classes):
-    out = torch.rand(batch, *output_shape)
-    module = postprocessor.ClassificationPostprocessor(output_shape, num_classes)
-    post_out = module(out)
-    assert list(post_out.shape) == [batch, num_classes]
-    assert not post_out.isnan().any()
-    post_out.mean().backward()
+@pytest.mark.parametrize('in_shape', [(1, 9), (3, 9)])
+@pytest.mark.parametrize('out_dim', [10])
+def test_classification_postprocessor(batch, in_shape, out_dim):
+    x = torch.rand(batch, *in_shape)
+    module = postprocessor.ClassificationPostprocessor(in_shape, out_dim)
+    out = module(x)
+    out_shape = [out_dim] if in_shape[0] == 1 else [in_shape[0], out_dim]
+    assert out_shape == module.out_shape
+    assert list(out.shape) == [batch, *module.out_shape]
+    assert not out.isnan().any()
+    out.mean().backward()
