+++ test/module/perceiver_io/test_preprocessor.py
@@ -13,20 +13,22 @@ def test_text_preprocessor(batch, vocab_size, embed_dim, max_seq_len):
     module = preprocessor.TextPreprocessor(vocab_size, embed_dim, max_seq_len)
     assert embed_dim == module.output_dim
     out = module(x)
-    assert list(out.shape) == [batch, max_seq_len, module.output_dim]
+    assert [max_seq_len, embed_dim] == module.out_shape
+    assert list(out.shape) == [batch, *module.out_shape]
     out.mean().backward()
 
 
 @pytest.mark.parametrize('batch', [2])
-@pytest.mark.parametrize('dim_shape', [
+@pytest.mark.parametrize('in_shape', [
     [64, 3],  # series
     [64, 64, 3],  # image
     [64, 64, 64, 3]  # volume
 ])
 @pytest.mark.parametrize('num_freq_bands', [32])
-def test_text_preprocessor(batch, dim_shape, num_freq_bands):
-    x = torch.rand(batch, *dim_shape)
-    module = preprocessor.FourierPreprocessor(dim_shape, num_freq_bands)
+def test_text_preprocessor(batch, in_shape, num_freq_bands):
+    x = torch.rand(batch, *in_shape)
+    max_reso = [2 * r for r in in_shape[:-1]]  # max resolution twice the input size
+    module = preprocessor.FourierPreprocessor(in_shape, num_freq_bands, max_reso)
     out = module(x)
-    flat_dim = math.prod(dim_shape[:-1])
-    assert list(out.shape) == [batch, flat_dim, module.output_dim]
+    assert [math.prod(in_shape[:-1]), module.out_dim] == module.out_shape
+    assert list(out.shape) == [batch, *module.out_shape]
