+++ torcharc/module/perceiver_io/postprocessor.py
@@ -1,28 +1,37 @@
 from torch import nn
 
 
+IdentityPostprocessor = nn.Identity
+
+
 class ProjectionPostprocessor(nn.Module):
-    '''Postprocessing to use a linear layer to project PerceiverDecoder output's last dimension, i.e. shape (O, E) -> (O, proj_dim).'''
+    '''Postprocessing to use a linear layer to project PerceiverDecoder output's last dimension, i.e. shape (O, E) -> (O, out_dim).'''
 
-    def __init__(self, output_shape: tuple, proj_dim: int):
+    def __init__(self, in_shape: list, out_dim: int):
         super().__init__()
-        output_o, output_e = output_shape  # PerceiverDecoder output shape (O, E)
-        self.proj_dim = proj_dim
-        self.linear = nn.Linear(output_e, self.proj_dim)
+        decoder_o, decoder_e = in_shape  # PerceiverDecoder output shape (O, E)
+        self.out_dim = out_dim
+        self.linear = nn.Linear(decoder_e, self.out_dim)
+        self.out_shape = [decoder_o, out_dim]
 
-    def forward(self, output):
-        return self.linear(output)
+    def forward(self, x):
+        return self.linear(x)
 
 
 class ClassificationPostprocessor(ProjectionPostprocessor):
-    '''Classification postprocessing for Perceiver. This is simply projecting decoder output to classes, i.e. shape (O, E) -> (num_classes)'''
+    '''
+    Classification postprocessing for Perceiver.
+    This is just ProjectionPostprocessor except it tries to squeeze dim=1 whenever possible, i.e. when PerceiverDecoder output shape:
+    - (O=1, E>=1): out_shape = [num_classes]; useful for plain classifier
+    - (O>1, E>=1): out_shape = [O, num_classes]; useful for sequence-classifier, i.e. O=max_seq_len, num_classes=vocab_size
+    '''
 
-    def __init__(self, output_shape: tuple, num_classes: int):
-        assert output_shape[0] == 1, 'output_shape[0] must be 1 for classification'
-        super().__init__(output_shape, num_classes)
+    def __init__(self, in_shape: list, out_dim: int):
+        super().__init__(in_shape, out_dim)
+        self.out_shape = [out_dim] if in_shape[0] == 1 else [in_shape[0], out_dim]
 
-    def forward(self, output):
-        return super().forward(output).squeeze(dim=1)  # output logits
+    def forward(self, x):
+        return super().forward(x).squeeze(dim=1)  # output logits
 
 # TODO embedding postprocessor
-# TODO multimodal postprocessor by partitioning output_shape[0] to each postprocessor
+# TODO multimodal postprocessor by partitioning in_shape[0] to each postprocessor
